refactor: move types from storage folder to navigation folder
