+++ src/app/navigation/RootNavigator.tsx
@@ -6,7 +6,7 @@ import { HomeLoading } from "@/app/screens/home/Home";
 import { useAuth } from "@/hooks/account/useAuth";
 import { i18nInitializedPromise } from "@/i18n";
 import { screenOptions } from "@/navigation/navigationConfig";
-import type { RootStackParamList } from "@/services/storage/types";
+import type { RootStackParamList } from "@/types";
 
 const Stack = createStackNavigator<RootStackParamList>();
