+++ src/app/screens/account/Account.tsx
@@ -18,7 +18,7 @@ import { TextSkeleton } from "@/components/Skeleton";
 import { QUERY_KEYS } from "@/constants";
 import { useTheme } from "@/contexts/ThemeContext";
 import { useUser } from "@/hooks/account/useUser";
-import type { AccountNavigation } from "@/services/storage/types";
+import type { AccountNavigation } from "@/types";
 import { getStudentYear } from "@/utils";
 
 export const Account = () => {
