+++ src/components/custom/card/HomeworkCard.tsx
@@ -9,7 +9,7 @@ import { TouchableOpacity, View } from "react-native";
 import { Text } from "@/components/common/Text";
 import { useTheme } from "@/contexts/ThemeContext";
 import type { Homework } from "@/dto";
-import type { AppStackParamList } from "@/services/storage/types";
+import type { AppStackParamList } from "@/types";
 
 interface Props {
   homework: Homework;
