+++ src/services/storage/types.ts
@@ -1,7 +1,3 @@
-import type { NavigatorScreenParams } from "@react-navigation/core";
-import type { StackNavigationProp } from "@react-navigation/stack";
-import type { Homework } from "@/dto";
-
 export type StorageItemValue = string | object | null;
 
 export interface StorageItem {
@@ -9,14 +5,6 @@ export interface StorageItem {
   timestamp: number;
   expiry?: number; // Time in milliseconds
 }
-
-// Auth Stack Types
-export type AuthStackParamList = {
-  Welcome: undefined;
-  Signin: undefined;
-  Signup: undefined;
-  ResetPassword: { email: string };
-};
 
 export interface StorageService {
   get<T>(key: string): Promise<T | null>;
@@ -35,71 +23,5 @@ export interface StorageService {
 
   multiRemove(keys: string[]): Promise<boolean>;
 }
-
-// Main App Stack Types
-export type AppStackParamList = {
-  Home: undefined;
-  WashingMachine: undefined;
-  Restaurant: undefined;
-  RestaurantReviews: { id: number };
-  Timetable: undefined;
-  Homework: undefined;
-  HomeworkDetails: { homework: Homework };
-  Games: undefined;
-  Profile: {
-    userId: string;
-  };
-  BottomTabNavigator: undefined;
-  Clubs: undefined;
-  Traq: undefined;
-  Olimtpe: undefined;
-  Account: undefined;
-};
-
-type AccountStackParamList = {
-  Account: undefined;
-  EditProfile: undefined;
-  Settings: undefined;
-  ChangePassword: undefined;
-  Feedback: undefined;
-};
-export type AccountNavigation = StackNavigationProp<AccountStackParamList>;
-
-type SettingsStackParamList = {
-  Account: undefined;
-  EditProfile: undefined;
-  ChangePassword: undefined;
-  Notifications: undefined;
-  Language: undefined;
-  Appearance: undefined;
-  About: undefined;
-  Help: undefined;
-  Statistics: undefined;
-  Legal: undefined;
-};
-export type SettingsNavigation = StackNavigationProp<SettingsStackParamList>;
-
-// Root Stack Types
-export type RootStackParamList = {
-  Auth: NavigatorScreenParams<AuthStackParamList>;
-  App: NavigatorScreenParams<AppStackParamList>;
-};
-
-// Navigation Types for useNavigation hook
-declare global {
-  namespace ReactNavigation {
-    interface RootParamList extends RootStackParamList {}
-  }
-}
-
-export const TAB_ROUTES = [
-  "Home",
-  "Services",
-  "Games",
-  "Account",
-  "Welcome",
-] as const;
-
-export type TabRoute = (typeof TAB_ROUTES)[number];
 
 export default StorageService;
