+++ src/types/navigation/navigation.types.ts
@@ -1,3 +1,7 @@
+import type { NavigatorScreenParams } from "@react-navigation/core";
+import type { StackNavigationProp } from "@react-navigation/stack";
+import type { Homework } from "@/dto";
+
 export type BottomTabParamList = {
   AccountScreen: undefined;
   HomeScreen: undefined;
@@ -6,3 +10,64 @@ export type BottomTabParamList = {
 };
 
 export default BottomTabParamList;
+
+export type AuthStackParamList = {
+  Welcome: undefined;
+  Signin: undefined;
+  Signup: undefined;
+  ResetPassword: { email: string };
+};
+
+export type AppStackParamList = {
+  Home: undefined;
+  WashingMachine: undefined;
+  Restaurant: undefined;
+  RestaurantReviews: { id: number };
+  Timetable: undefined;
+  Homework: undefined;
+  HomeworkDetails: { homework: Homework };
+  Games: undefined;
+  Profile: {
+    userId: string;
+  };
+  BottomTabNavigator: undefined;
+  Clubs: undefined;
+  Traq: undefined;
+  Olimtpe: undefined;
+  Account: undefined;
+};
+
+type AccountStackParamList = {
+  Account: undefined;
+  EditProfile: undefined;
+  Settings: undefined;
+  ChangePassword: undefined;
+  Feedback: undefined;
+};
+export type AccountNavigation = StackNavigationProp<AccountStackParamList>;
+
+type SettingsStackParamList = {
+  Account: undefined;
+  EditProfile: undefined;
+  ChangePassword: undefined;
+  Notifications: undefined;
+  Language: undefined;
+  Appearance: undefined;
+  About: undefined;
+  Help: undefined;
+  Statistics: undefined;
+  Legal: undefined;
+};
+export type SettingsNavigation = StackNavigationProp<SettingsStackParamList>;
+
+export type RootStackParamList = {
+  Auth: NavigatorScreenParams<AuthStackParamList>;
+  App: NavigatorScreenParams<AppStackParamList>;
+};
+
+// Navigation Types for useNavigation hook
+declare global {
+  namespace ReactNavigation {
+    interface RootParamList extends RootStackParamList {}
+  }
+}
